requirements.txt
# PDF handling
fpdf2>=2.7.0
pdfplumber>=0.10.0

# Vector store
chromadb>=0.4.0

# Graph
networkx>=3.0

# Embeddings — local, no API needed
sentence-transformers>=2.2.0

# LLM — via Ollama local server
ollama>=0.3.0

# Utilities
rich>=13.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
